gpu_requirements.txt
absl-py==0.7.1
astor==0.8.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
cycler==0.10.0
dill==0.3.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.3.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.4
mock==3.0.5
multiprocessing==2.6.2.1
numpy==1.16.4
pandas==0.24.2
protobuf==3.9.0
pydicom==1.3.0
pydot==1.4.1
pygamma==4.3.3
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scipy==1.2.2
seaborn==0.9.0
six==1.12.0
sobol-seq==0.1.2
soupsieve==1.9.2
subprocess32==3.5.4
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow-gpu==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.32.2
virtualenv==16.6.2
Werkzeug==0.15.5
wrapt==1.11.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
